ESYNC-3753 populate x-xl4bus.timestamp with correct timestamp value
We were populating this property with the address of the pf_sec_time
function rather than calling the pf_sec_time function and using the
result to populate the "timestamp" property value.

With this change, the actual date/time is represented (in seconds
sine Jan-1-1970) in this property value, as can be verified by comparing
against the current/recent time() value (or 'date +%s').